Move query tests to PoolPriceOracle
