+++ pkg/pool-utils/contracts/test/oracle/MockPoolPriceOracle.sol
@@ -31,6 +31,16 @@ contract MockPoolPriceOracle is MockSamples, PoolPriceOracle {
 
     event PriceDataProcessed(bool newSample, uint256 sampleIndex);
 
+    uint256 private _mockedOracleIndex;
+
+    function mockOracleIndex(uint256 index) external {
+        _mockedOracleIndex = index;
+    }
+
+    function _getOracleIndex() internal view virtual override returns (uint256) {
+        return _mockedOracleIndex;
+    }
+
     function mockSample(uint256 index, Sample memory sample) public {
         _samples[index] = encode(sample);
     }
@@ -77,9 +87,5 @@ contract MockPoolPriceOracle is MockSamples, PoolPriceOracle {
         uint256 timestamp
     ) external view returns (int256) {
         return _getPastAccumulator(variable, currentIndex, block.timestamp - timestamp);
-    }
-
-    function _getOracleIndex() internal view virtual override returns (uint256) {
-        return 0;
     }
 }
